feat: add rabbitMQ module
add exchangeType 'fanout'
because there are problems with receiving user registration event for 2 services:
notification and user.

It was decided to send events not to a specific 1 queue,
but to one exchange. In our case - to "fanout".
It allows you to send the event to 1 exchange,
which is linked to the required queues (bindings).

Fanout exchangeType does not send routingKey,
so it was decided to add "eventType" to published events.

!!! Note:
For correct operation you need to manually configure queues
and connect them to the exchange in the RabbitMQ interface