feat: add new client to the auth module for events
Now there are 2 clients: 1 is responsible for sending messages to the queue
(for receiving 1 message by 2 services)
and 1 is responsible for sending messages to 1 AUTH_QUEUE queue.
To fanout a client you should use the "publishToExchange" function from rabbitMQService.
For normal client you can just "emit" using routingKey.
Also:
- added a role to the event when email changes
for the corresponding subsequent redirect (in notification-service)
- changed the response data for password reset,
since the user must have the same data as when logging in