feat: rename verification code to code
now after resetting the password at the stage of sending the code to email,
the token is sent to the front end and the user is already authorized with the token.

Also corrected data when publishing an event.
Now always in events with code its type is sent in dto