feat: change the response data for password recovery, verify and resend code
now the user receives the token and user data upon successful verification of the code,
not upon requesting a password.
Since the previous technique of receiving data by email was insecure.

also now the update function of the verification code performs the upsert function,
i.e. if there is no such code, a new code is created in the database